+++ src/chart/speclayout/bubblechart_impl.h
@@ -15,8 +15,7 @@ namespace Charts
 class BubbleChartImpl : public BubbleChart
 {
 public:
-	BubbleChartImpl(const std::vector<double> &sizes,
-	    Boundary boundary,
+	explicit BubbleChartImpl(const std::vector<double> &sizes,
 	    const Geom::Rect &rect = Geom::Rect(Geom::Point(0, 0),
 	        Geom::Size(1, 1)));
 
@@ -29,15 +28,12 @@ private:
 
 	std::vector<RadiusRecord> radiuses;
 
-	Boundary boundary;
-	Geom::Point center;
-	double allSize;
-
 	void generate();
-	void addCircle(size_t index, const Geom::Circle &circle);
-	void updateCenter(const Geom::Circle &circle);
-	double sqrCenterDistance(const Geom::Point &p) const;
-	bool overlapsAny(const Geom::Circle &circle) const;
+	
+	std::optional<Geom::Circle> getTouchingCircle(
+		const RadiusRecord &act, 
+		size_t firstIdx, 
+		size_t lastIdx);
 };
 
 }